moved helper to core
