+++ src/core/containers/error-simulation/SimulateSyncError.js
@@ -3,8 +3,7 @@ import { compose } from 'redux';
 import React from 'react';
 
 import log from 'core/logger';
-
-import { render404WhenNotAllowed } from './utils';
+import { render404WhenNotAllowed } from 'core/utils';
 
 export class SimulateSyncErrorBase extends React.Component {
   render() {
